# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HireHand API",
    "version": "1.0.0",
    "description": "Minimal API for auth and jobs"
  },
  "servers": [
    { "url": "http://localhost:4000" }
  ],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "nullable": true },
          "email": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "location": { "type": "string", "nullable": true },
          "budget": { "type": "number", "nullable": true },
          "postedById": { "type": "integer", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "JobCreate": {
        "type": "object",
        "required": ["title", "description"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "location": { "type": "string" },
          "budget": { "type": "number" }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "type": "object", "properties": { "user": { "$ref": "#/components/schemas/User" } } } } } },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login and receive access token + httpOnly refresh cookie",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "accessToken": { "type": "string" }, "user": { "$ref": "#/components/schemas/User" } } } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Refresh access token using httpOnly cookie",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "accessToken": { "type": "string" }, "user": { "$ref": "#/components/schemas/User" } } } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout and clear refresh cookie",
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/jobs": {
      "get": {
        "summary": "List jobs",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "jobs": { "type": "array", "items": { "$ref": "#/components/schemas/Job" } } } } } } }
        }
      },
      "post": {
        "summary": "Create a job (requires Authorization: Bearer <token>)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/JobCreate" } }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "type": "object", "properties": { "job": { "$ref": "#/components/schemas/Job" } } } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}